backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yfinance==0.2.36
python-dotenv==1.0.0
openai==1.12.0
httpx==0.26.0
